app/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==4.10.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
